+++ source/main/webpack.config.js
@@ -1,19 +1,27 @@
 const webpack = require('webpack');
-const HardSourceWebpackPlugin = require('hard-source-webpack-plugin');
-const lodash = require('lodash');
 const yamljs = require('yamljs');
+// const HardSourceWebpackPlugin = require('hard-source-webpack-plugin');
+// const lodash = require('lodash');
 
 let reportUrl = '';
 reportUrl = yamljs.parseFile('launcher-config.yaml').reportServer;
 
 // Process env flags from buildkite and appveyor
-const isCi = process.env.CI && process.env.CI !== '';
+// const isCi = process.env.CI && process.env.CI !== '';
 
 module.exports = {
+  mode: 'development',
   devtool: 'cheap-module-source-map',
-  entry: './source/main/index.js',
+  entry: {
+    index: './source/main/index.js',
+    preload: './source/main/preload.js',
+  },
+  optimization: {
+    // https://github.com/webpack/webpack/issues/7470
+    nodeEnv: false,
+  },
   output: {
-    filename: 'index.js'
+    filename: '[name].js',
   },
   /**
    * Set targed to Electron speciffic node.js env.
@@ -40,6 +48,23 @@ module.exports = {
           loader: 'babel-loader',
         },
       },
+      {
+        test: /(pdfkit|linebreak|fontkit|unicode|brotli|png-js).*\.js$/,
+        use: {
+          loader: 'transform-loader?brfs',
+        }
+      },
+      {
+        test: /\.(woff2?|eot|ttf|otf|png|jpe?g|gif|svg)(\?.*)?$/,
+        exclude: /\.inline\.svg$/,
+        use: {
+          loader: 'file-loader',
+          options: {
+            name: '[name]-[hash].[ext]',
+            outputPath: 'assets/'
+          }
+        }
+      },
     ]
   },
   plugins: [
@@ -49,20 +74,23 @@ module.exports = {
       'process.env.MOBX_DEV_TOOLS': process.env.MOBX_DEV_TOOLS || 0,
       'process.env.BUILD_NUMBER': JSON.stringify(process.env.BUILD_NUMBER || 'dev'),
       'process.env.REPORT_URL': JSON.stringify(reportUrl),
+      'process.env.IS_WATCH_MODE': process.env.IS_WATCH_MODE === 'true'
     }, process.env.NODE_ENV === 'production' ? {
       // Only bake in NODE_ENV value for production builds.
       'process.env.NODE_ENV': '"production"',
     } : {})),
-    !isCi && (
-      new HardSourceWebpackPlugin({
-        configHash: (webpackConfig) => (
-          // Remove the `watch` flag to avoid different caches for static and incremental builds
-          require('node-object-hash')({ sort: false }).hash(lodash.omit(webpackConfig, 'watch'))
-        ),
-        environmentPaths: {
-          files: ['.babelrc', 'yarn.lock'],
-        },
-      })
-    )
+    // Hard source plugin is broken for webpack 4 :(
+    // https://github.com/mzgoddard/hard-source-webpack-plugin/issues/443
+    // !isCi && (
+    //   new HardSourceWebpackPlugin({
+    //     configHash: (webpackConfig) => (
+    //       // Remove the `watch` flag to avoid different caches for static and incremental builds
+    //       require('node-object-hash')({ sort: false }).hash(lodash.omit(webpackConfig, 'watch'))
+    //     ),
+    //     environmentPaths: {
+    //       files: ['.babelrc', 'yarn.lock'],
+    //     },
+    //   })
+    // )
   ].filter(Boolean),
 };