+++ source/renderer/app/App.js
@@ -1,5 +1,5 @@
 // @flow
-import React, { Component } from 'react';
+import React, { Component, Fragment } from 'react';
 import { Provider, observer } from 'mobx-react';
 import { ThemeProvider } from 'react-polymorph/lib/components/ThemeProvider';
 import DevTools from 'mobx-react-devtools';
@@ -7,7 +7,7 @@ import { Router } from 'react-router';
 import { IntlProvider } from 'react-intl';
 import { Routes } from './Routes';
 import { daedalusTheme } from './themes/daedalus';
-import environment from '../../common/environment';
+import { themeOverrides } from './themes/overrides/index.js';
 import translations from './i18n/translations';
 import type { StoresMap } from './stores/index';
 import type { ActionsMap } from './actions/index';
@@ -20,29 +20,32 @@ export default class App extends Component<{
   actions: ActionsMap,
   history: Object,
 }> {
+  componentWillMount() {
+    // loads app's global environment variables into AppStore via ipc
+    this.props.actions.app.initAppEnvironment.trigger();
+  }
   render() {
     const { stores, actions, history } = this.props;
     const { app } = stores;
     const locale = stores.profile.currentLocale;
-    const mobxDevTools = environment.MOBX_DEV_TOOLS ? <DevTools /> : null;
+    const mobxDevTools = global.environment.mobxDevTools ? <DevTools /> : null;
     const currentTheme = stores.profile.currentTheme;
-    const themeVars = require(`./themes/daedalus/${currentTheme}.js`); // eslint-disable-line
-
+    const themeVars = require(`./themes/daedalus/${currentTheme}.js`).default;
     return (
-      <div>
+      <Fragment>
         <ThemeManager variables={themeVars} />
         <Provider stores={stores} actions={actions}>
-          <ThemeProvider theme={daedalusTheme}>
+          <ThemeProvider theme={daedalusTheme} themeOverrides={themeOverrides}>
             <IntlProvider {...{ locale, key: locale, messages: translations[locale] }}>
-              <div style={{ height: '100%' }}>
+              <Fragment>
                 <Router history={history} routes={Routes} />
                 {mobxDevTools}
                 {app.isAboutDialogOpen && <AboutDialog />}
-              </div>
+              </Fragment>
             </IntlProvider>
           </ThemeProvider>
         </Provider>
-      </div>
+      </Fragment>
     );
   }
 }