+++ source/renderer/app/components/status/NetworkStatus.js
@@ -18,15 +18,18 @@ import { UNSYNCED_BLOCKS_ALLOWED } from '../../config/numbersConfig';
 import { getNetworkEkgUrl } from '../../utils/network';
 import closeCross from '../../assets/images/close-cross.inline.svg';
 import LocalizableError from '../../i18n/LocalizableError';
-import { CardanoNodeStates } from '../../../../common/types/cardanoNode.types';
-import environment from '../../../../common/environment';
+import { CardanoNodeStates } from '../../../../common/types/cardano-node.types';
 import styles from './NetworkStatus.scss';
-import type { CardanoNodeState } from '../../../../common/types/cardanoNode.types';
+import type { CardanoNodeState } from '../../../../common/types/cardano-node.types';
 
 let syncingInterval = null;
 
 type Props = {
   cardanoNodeState: ?CardanoNodeState,
+  isDev: boolean,
+  isMainnet: boolean,
+  isStaging: boolean,
+  isTestnet: boolean,
   isNodeResponding: boolean,
   isNodeSubscribed: boolean,
   isNodeSyncing: boolean,
@@ -115,9 +118,9 @@ export default class NetworkStatus extends Component<Props, State> {
 
   render() {
     const {
-      cardanoNodeState, isNodeResponding, isNodeSubscribed, isNodeSyncing, isNodeInSync,
-      isNodeTimeCorrect, isConnected, isSynced, syncPercentage, hasBeenConnected,
-      localTimeDifference, isSystemTimeCorrect, isForceCheckingNodeTime,
+      cardanoNodeState, isDev, isTestnet, isStaging, isMainnet, isNodeResponding, isNodeSubscribed,
+      isNodeSyncing, isNodeInSync, isNodeTimeCorrect, isConnected, isSynced, syncPercentage,
+      hasBeenConnected, localTimeDifference, isSystemTimeCorrect, isForceCheckingNodeTime,
       mostRecentBlockTimestamp, localBlockHeight, networkBlockHeight,
       onForceCheckLocalTimeDifference, onClose, nodeConnectionError, isSystemTimeIgnored,
       onOpenExternalLink,
@@ -149,7 +152,9 @@ export default class NetworkStatus extends Component<Props, State> {
     ]);
 
     // Cardano Node EKG server is not enabled for the Mainnet!
-    const cardanoNodeEkgLink = environment.isMainnet() ? false : getNetworkEkgUrl();
+    const cardanoNodeEkgLink = isMainnet ? false : getNetworkEkgUrl({
+      isDev, isStaging, isTestnet
+    });
 
     return (
       <div className={styles.component}>