+++ source/renderer/app/containers/wallet/PaperWalletCreateCertificatePage.js
@@ -47,8 +47,9 @@ export default class PaperWalletCreateCertificatePage extends Component<Props, S
   };
 
   render() {
-    const { uiDialogs } = this.props.stores;
+    const { uiDialogs, app } = this.props.stores;
     const { showConfirmationDialog } = this.state;
+    const { environment: network } = app;
     let activeDialog = null;
 
     if (uiDialogs.isOpen(InstructionsDialog)) {
@@ -100,6 +101,7 @@ export default class PaperWalletCreateCertificatePage extends Component<Props, S
         {activeDialog}
         {showConfirmationDialog && (
           <ConfirmationDialog
+            network={network}
             onCancel={this.hideConfirmationDialog}
             onConfirm={this.onClose}
           />