+++ storybook/stories/index.js
@@ -15,5 +15,4 @@ import './TopBar.stories';
 import './WalletTransactionsList.stories';
 import './ExportWalletToFileDialog.stories';
 import './PaperWallets.stories';
-import './AntivirusRestaurationSlowdownNotification.stories';
 import './WalletScreens.stories';